+++ pkg/testfixtures/storage/storage.go
@@ -66,6 +66,8 @@ func RunDatastoreTestContainer(t testing.TB, engine string) DatastoreTestContain
 		return memoryTestContainer{}
 	case "sqlite":
 		return NewSqliteTestContainer().RunSqliteTestDatabase(t)
+	case "dsql":
+		return NewDSQLTestContainer().RunDSQLTestContainer(t)
 	default:
 		t.Fatalf("unsupported datastore engine: %q", engine)
 		return nil